+++ src/main/java/com/github/johnmedlockdev/main/Main.java
@@ -13,31 +13,31 @@ public class Main {
     public static void main(String... args) throws SQLException, IOException, ClassNotFoundException {
 
         // TODO: I'd like to refactor all this into a one liner. ^ deal with errors with proper blocks.
-        while (true) {
-            ParseInput userInput = new ParseInput(args);
+//        while (true) {
+        ParseInput userInput = new ParseInput(args);
 
-            if (userInput.getMode().equals("")) {
+        if (userInput.getMode().equals("")) {
 
-                FileInfo fileInfo = new FileInfo(userInput);
-                Data data = new Data(userInput, fileInfo);
+            FileInfo fileInfo = new FileInfo(userInput);
+            Data data = new Data(userInput, fileInfo);
 
-                data.logic();
+            data.logic();
 
-            } else {
-                // need to refactor this
-                if (userInput.getMode().equals("exit")) {
-                    break;
-                    //TODO: this needs to properly exit the loop
-                }
-                System.out.println("What would you like to do with the database?");
-                Scanner scanner = new Scanner(System.in);  // Create a Scanner object
-                String function = scanner.nextLine();  // Read user input
-                Database dataDB = new Database(function);
+        } else {
+            // need to refactor this
+            if (userInput.getMode().equals("exit")) {
+//                    break; // commented out while loop
+                //TODO: this needs to properly exit the loop
             }
-
+            System.out.println("What would you like to do with the database?");
+            Scanner scanner = new Scanner(System.in);  // Create a Scanner object
+            String function = scanner.nextLine();  // Read user input
+            Database dataDB = new Database(function);
         }
+
     }
 }
+//}
 
 // Desired output:
 // 1. feed in SPY data